requirements.txt
streamlit
requests
firebase-admin
gTTS
google-cloud-firestore